+++ src/components/PledgePackage/ListPledgePackages/index.js
@@ -1,20 +1,13 @@
 import React, { useContext, useEffect, useState } from 'react';
 import { SectionInner } from '../../Layout/Sections';
 import * as s from './style.module.less';
-import AvatarImage from '../../AvatarImage';
-import {
-  useGetMostRecentInteractions,
-  useUpdateInteraction,
-} from '../../../hooks/Api/Interactions';
+import { useGetMostRecentInteractions } from '../../../hooks/Api/Interactions';
 
 import { CTALink } from '../../Layout/CTAButton';
 import AuthContext from '../../../context/Authentication';
 
-import paketSvg from '../paket-v2.svg';
-import check from '../check.svg';
-
 import { LoadingAnimation } from '../../LoadingAnimation';
-import cN from 'classnames';
+import { Package } from './Package';
 
 export default () => {
   const [state, pledgePackages, getInteractions] =
@@ -47,8 +40,15 @@ export default () => {
 
   return (
     <SectionInner wide={true}>
-      <h2 className={s.headingViolet}>Alle Sammelpakete</h2>
-      <p>Intro-Text zu Sammelpaketen</p>
+      <h2 className={s.violet}>Alle Sammelpakete</h2>
+      <p>
+        Zeig deinen Einsatz für's Grundeinkommen und setze dir ein Sammelziel!
+        Es gibt Pakete mit jeweils einem Ziel von 50 Unterschriften, von denen
+        du dir so viele nehmen kannst, wie du möchtest!{' '}
+        {userData.interactions &&
+          packagesOfUser.length === 0 &&
+          'Mach mit und schnapp dir dein erstes Paket!'}
+      </p>
 
       {state && state !== 'loading' && (
         <p>
@@ -67,7 +67,7 @@ export default () => {
 
       {packagesOfUser.length > 0 && (
         <div>
-          <h3 className={s.headingViolet}>Deine Pakete</h3>
+          <h3 className={s.violet}>Deine Pakete</h3>
           <p>
             Du hast dir {packagesOfUser.length} Pakete geschnappt und somit
             versprochen, {packagesOfUser.length * 50} Unterschriften zu sammeln.
@@ -103,11 +103,11 @@ export default () => {
       {state && state !== 'loading' ? (
         <>
           {pledgePackages.length > 0 ? (
-            <h3 className={s.headingViolet}>
+            <h3 className={s.violet}>
               Diese Pakete hat sich schon jemand geschnappt
             </h3>
           ) : (
-            <h3 className={s.headingViolet}>Noch keine Pakete verteilt!</h3>
+            <h3 className={s.violet}>Noch keine Pakete verteilt!</h3>
           )}
 
           <div className={s.container}>
@@ -125,9 +125,7 @@ export default () => {
 
           {pledgePackagesDone.length > 0 && (
             <>
-              <h3 className={s.headingViolet}>
-                Diese Pakete wurden schon erledigt!
-              </h3>
+              <h3 className={s.violet}>Diese Pakete wurden schon erledigt!</h3>
               <div className={s.container}>
                 {pledgePackagesDone.map((pledgePackage, index) => {
                   return (
@@ -149,123 +147,4 @@ export default () => {
       )}
     </SectionInner>
   );
-};
-
-const Package = ({
-  body,
-  user,
-  createdAt,
-  id,
-  done,
-  belongsToCurrentUser = false,
-  showDone = false,
-}) => {
-  const [pledgeUpdateState, updatePledgePackage] = useUpdateInteraction();
-  const { updateCustomUserData } = useContext(AuthContext);
-  const [, , getInteractions] = useGetMostRecentInteractions();
-
-  useEffect(() => {
-    if (pledgeUpdateState === 'saved') {
-      getInteractions(null, 0, 'pledgePackage');
-      updateCustomUserData();
-    }
-  }, [pledgeUpdateState]);
-
-  return (
-    <div
-      className={cN(s.fullPackage, {
-        [s.extraBottomMargin]: belongsToCurrentUser || showDone,
-      })}
-    >
-      <div className={s.packageIconContainer}>
-        <img
-          src={paketSvg}
-          className={s.packageIcon}
-          alt="Symbolbild eines Paketes"
-        />
-        <AvatarImage className={s.avatar} user={user} sizes="120px" />
-        {belongsToCurrentUser ? (
-          <>
-            {pledgeUpdateState === 'saving' ? (
-              <div className={s.loadingPackageUpdate}>
-                <LoadingAnimation />
-              </div>
-            ) : (
-              <>
-                {!done && pledgeUpdateState !== 'saved' ? (
-                  <button
-                    onClick={() =>
-                      updatePledgePackage({
-                        id: id,
-                        done: true,
-                      })
-                    }
-                    className={cN(
-                      s.linkLikeFormattedButton,
-                      s.onWhiteBackground,
-                      s.setDone
-                    )}
-                  >
-                    <b>Als erledigt markieren</b>
-                  </button>
-                ) : (
-                  <p className={s.isDone}>
-                    <img
-                      src={check}
-                      className={s.checkIcon}
-                      alt="Häkchen-Icon"
-                    />
-                    <b>ERLEDIGT!</b>
-                  </p>
-                )}
-              </>
-            )}
-          </>
-        ) : (
-          <>
-            {showDone && (
-              <p className={s.isDone}>
-                <img src={check} className={s.checkIcon} alt="Häkchen-Icon" />
-                <b>ERLEDIGT!</b>
-              </p>
-            )}
-          </>
-        )}
-      </div>
-      <div className={s.packageTextContainer}>
-        <h4 className={s.name}>{user.username}</h4>
-        <p className={s.createdAt}>Vor {getElapsedTime(createdAt)}</p>
-        <p className={s.quote}>"{body}"</p>
-      </div>
-    </div>
-  );
-};
-
-const getElapsedTime = createdAt => {
-  const endTime = new Date();
-  const startTime = new Date(createdAt);
-  const timeDiff = endTime.getTime() - startTime.getTime();
-  const seconds = Math.floor(timeDiff / 1000);
-  const minutes = Math.floor(seconds / 60);
-  const hours = Math.floor(minutes / 60);
-  const days = Math.floor(hours / 24);
-  const weeks = Math.floor(days / 7);
-  const months = Math.floor(weeks / 4);
-  const years = Math.floor(months / 12);
-
-  if (years > 0) {
-    return `${years} ${years === 1 ? 'Jahr' : 'Jahre'}`;
-  } else if (months > 0) {
-    return `${months} ${months === 1 ? 'Monat' : 'Monate'}`;
-  } else if (weeks > 0) {
-    return `${weeks} ${weeks === 1 ? 'Woche' : 'Wochen'}`;
-  } else if (days > 0) {
-    return `${days} ${days === 1 ? 'Tag' : 'Tage'}`;
-  } else if (hours > 0) {
-    return `${hours} ${hours === 1 ? 'Stunde' : 'Stunden'}`;
-  } else if (minutes > 0) {
-    return `${minutes} ${minutes === 1 ? 'Minute' : 'Minuten'}`;
-  } else if (seconds > 0) {
-    return `${seconds} ${seconds === 1 ? 'Sekunde' : 'Sekunden'}`;
-  }
 };
